+++ client/components/SingleWineCartQty.js
@@ -1,0 +1,48 @@
+import React from "react";
+import { connect } from "react-redux";
+import { fetchCart, checkoutAll } from "../store/cart";
+import { Link } from "react-router-dom";
+
+class SingleWineCartQty extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      cartItemQty: 1,
+    };
+    this.handleChange = this.handleChange.bind(this);
+    this.createDropDown = this.createDropDown.bind(this);
+  }
+
+  handleChange(event) {
+    this.setState({
+      [event.target.name]: event.target.value,
+    });
+  }
+
+  createDropDown(stockQty) {
+    let elements = [];
+    for (let i = 1; i <= stockQty; i++) {
+      elements.push(<option key={i}>{i}</option>);
+    }
+    return elements;
+  }
+
+  render() {
+    const singleItemInCart = this.props.cartItem;
+    // console.log("******************", singleItemInCart);
+    return (
+      <div className="cart-item-qty-dropdown-container">
+        <select
+          className="qty-dropdown"
+          name="cartItemQty"
+          onChange={this.handleChange}
+          value={this.state.cartItemQty}
+        >
+          {this.createDropDown(singleItemInCart.quantity)}
+        </select>
+      </div>
+    );
+  }
+}
+
+export default SingleWineCartQty;
